+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Binder for Pandora's notebooks. [#215] 
 - Version handling with setuptools_scm [#212]
 - Set dataset's transform to None if it is the identity matrix. [#211]
+- Handling of np.inf values on input images. [#210]
 
 ### Fixed